requirements.txt
numpy~=1.24.4
opencv_python==4.9.0.80
pyaudio~=0.2.11
PyQt6~=6.4.2
qdarkstyle==3.2.3
qtpy==2.4.1
torch==2.4.0
torchvision==0.19.0
vosk==0.3.45

pillow~=10.4.0
pytesseract~=0.3.13